Add slack obfuscated URL
The atProvider.obfuscatedUrl property is used to store the obfuscated url returned by the API. Since Dynatrace handles the URL as a secret (since it contains secret information), the API response is obfuscated instead of having the plain URL in place. Without this change, Crossplane would detect the obfuscated URL as changed URL. To mitigate this issue, we store the obfuscated url and compare that one instead of the plain url.